fixes #51: update size length
fixes #51: Refactor model folder
